Fix Magic Wand

+++ code/MImage.h
@@ -34,7 +34,7 @@ class MImage{
 		/* Constructors/destructor */
 		MImage(void); /* default constructor */
 		MImage(int xs,int ys,int zs); 
-		MImage(int xs, int ys, int zs, int color);
+		MImage(int xs, int ys, int zs, float color);
 		MImage(const MImage &copy); /* copy constructor */ 
 		~MImage(void);
